change name of center_rv() to center()
This is a nominal change only; however, the idea is to make the verb more expansive in its scope. Rather than just being for running variables in an RDD, this will center any variable over any given center-point. 

The idea is that it should be useful for mean-centering in MLM/HLM for variables beyond just rvs, as in:

```
df %>% 
   center(var, mean(var))
```